[SERVER] modify | impl handling exception anonymousUser, change jpaAudit return type

+++ intagralServer/src/main/java/com/a304/intagral/config/JpaAuditConfig.java
@@ -11,7 +11,7 @@ import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
 public class JpaAuditConfig {
 
     @Bean
-    public AuditorAware<Long> auditorProvider() {
+    public AuditorAware<Integer> auditorProvider() {
         return new AuditorAwareImpl();
     }
 }
